Use C++ lambdas to simplify instruction tests.
This is a first pass to remove boiler plate from the tests. There is more that
can be done likely.

Also, the direction taken here is to over-specify the state (eg setting up the
CompiledCode and MachineCode objects for every test). The advantage of this
approach is that the execution of the instructions are less likely to result
in confusing exceptions and memory faults. The overhead in C++ should be
negligible.