Disallow fork() from non-root Fiber. Related to #3725.
In Rubinius, Fiber is composed with a native OS thread. A consequence of this
is that when fork() is called, *only* the current OS thread is carried over to
the new process.

This contrasts with implementing Fiber as a user-mode "stack" that is
multiplexed onto a single OS thread. In this case, all the data from multiple
Fibers are carried over to the new process with the OS thread executing the
fork() system call.

The user-mode "stack" implementation has a significant drawback in not being
reliably portable across systems (eg porting to WebAssembly would not permit
physically modifying a "thread" stack). Additionally, this approach obscures
the program's operation from the OS scheduler and complicates debugging the
process.

The advantages of composing Fiber with a first-class OS thread are the
opposite. They are completely portable, they offer the OS clear visibility
into the operation of the program, they inspectable by debuggers, and the
resource overhead is as readily controllable as with user-mode "stack" Fibers.

The drawbacks of the user-mode "stack" approach are significant, so Rubinius
has chosen to implement Fiber using first-class OS threads. This makes use of
Fiber incompatible with using fork(). However, Fiber is nothing more than a
limited Thread, and since Rubinius gives full control over the resources
allocated for Thread, Fiber offers no significant advantage over Thread aside
from the abstraction of the single lock associated with all the Fibers on a
Thread that guarantees serial execution of all Fiber code. This could be
mimicked by pure Ruby code and instances of Thread.

Since the use case of Fiber and fork() are mostly distinct, the limitation of
not being able to call fork() from a non-root Fiber is considered an
acceptable trade-off for the advantages of proper OS thread support of Fiber.