Cleaned up Thread init_lock.
At one time, starting a Thread used a lock and this naming was a relic of
that. Now creating a Thread is properly sequenced and does not need a lock.

Also, this lock is used at execution leaf nodes around state and when no
fork() can be active, so re-init'ing the lock on process transitions is not
necessary.