Simplify MachineThreads.
MachineThreads was introduced when state control around fork/exec was ad hoc
and scattered. The boundaries were also not well-defined and there were
numerous race conditions.

The boundaries are now much better defined and some of the 'events' have not
been needed. There is more restructuring needed and if the removed events turn
out to be required, they can be re-added.