Fixed inline cache for recursive calls.
We cache a method at a call site _after_ the method has been invoked so
that we don't cache methods that raised exceptions.

But when a method is called recursively, the state of the cache will
have changed and when unwinding from the dispatch call, we need check
that the state of the cache is what we assume it to be. A similar thing
can happen when code is executing concurrently.