Added --gc-stack-check.
When the process is going to checkpoint to run GC, all the threads will
log their stacks.

Presently, if a thread is already suspended, it won't log its stack.
When the GC triggering and checkpointing is reworked, the threads that
were suspended when GC occurred will check a flag that GC has run when
resuming to log their stacks.