Hold logger lock during fork().
A thread that does not stop for fork() because it's not mutating memory
may still use the logger, which causes the logger lock to be in an
indeterminate state across a fork() call, possibly resulting in the lock
being locked in the child. By explicitly holding the lock, we can safely
unlock it in the child and the lock is in a determinate state.