Added call instruction variants.
The call_send instruction looks for a function and falls back to
dispatching a method. This will replace the send_stack instruction that
is used for implicit self sends in code that could have functions.

The call instruction only looks for a function.

The call_0 instruction is for 0-arity functions.

The send_vcall instruction will be modified to look for functions.